Move IncludeStoryOption.tsx file. Change styling of checkbox.
