+++ lib/ReactViews/Map/Panels/SharePanel/share-panel.scss
@@ -162,3 +162,8 @@
     cursor: pointer;
   }
 }
+
+.checkbox {
+  margin-top: 0.5em;
+  float: right;
+}
