+++ lib/ReactViews/Map/Panels/SharePanel/share-panel.scss.d.ts
@@ -13,6 +13,7 @@ interface CssExports {
   'catalog-share-inner': string;
   'catalogShare': string;
   'catalogShareInner': string;
+  'checkbox': string;
   'clipboard-for-catalog-share': string;
   'clipboard-for-story-share': string;
   'clipboardForCatalogShare': string;
